+++ bin/install-local-development-env.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+git clone https://github.com/WordPress/wordpress-develop.git vendor/wordpress-develop;