+++ source/Themes/AbstractTheme.php
@@ -17,6 +17,22 @@ abstract class AbstractTheme implements ThemeInterface
     /**
      * @inheritdoc
      */
+    public function run()
+    {
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function runAdmin()
+    {
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function getContainer()
     {
         return $this->container;