+++ tests/AlmostControllers/HttpStackTest.php
@@ -20,7 +20,7 @@ class HttpStackTest extends \WP_UnitTestCase
 
     public function testRegister()
     {
-        $this->setExpectedException(\LogicException::class);
+        $this->setExpectedException(\LogicException::class, 'You need set actions before call register method.');
         $this->stub->register();
 
         $this->setExpectedException(null);