+++ tests/Plugins/AbstractPluginTest.php
@@ -23,6 +23,16 @@ class AbstractPluginTest extends \WP_UnitTestCase
         );
     }
 
+    public function testRun()
+    {
+        $this->assertSame($this->stub, $this->stub->run());
+    }
+
+    public function testRunAdmin()
+    {
+        $this->assertSame($this->stub, $this->stub->runAdmin());
+    }
+
     public function testGetFile()
     {
         $this->assertSame(__FILE__, $this->stub->getFile());