+++ tests/TermMeta/AbstractTermMetaTest.php
@@ -436,7 +436,7 @@ class AbstractTermMetaTest extends \WP_UnitTestCase
             )
         );
 
-        $this->setExpectedException(\Exception::class);
+        $this->setExpectedException(\Exception::class, 'Expected argument of type "array", "string" given');
         $this->stub->validate();
     }