+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.10.1 - October 19, 2018
+
+* Added runAdmin method for plugins and themes ([#36](https://github.com/korobochkin/wp-kit/issues/36)).
+* Updated repository readme.
+
 ## 0.10.0 - May 14, 2018
 
 * Added `NodeFactory` class for building Options, Transients, Post Meta, Term Meta ([#29](https://github.com/korobochkin/wp-kit/issues/29)).